+++ gametime-api/src/main/java/com/gametime/api/TeamService.java
@@ -2,13 +2,17 @@ package com.gametime.api;
 
 import com.gametime.common.Team;
 
+import java.util.List;
+
 /**
  * Created by dave on 2/6/2017.
  */
 public interface TeamService {
-    Team getTeam(Long teamId);
-    /*
+    Team getTeam(Integer teamId);
     List<Team> getAllTeams();
+
+    /*
+
     List<Team> getTeamsByConference(int conferenceId);
     List<Team> getTeamsByDivision(int divisionId);
     */
